Fixed backend

+++ Repositories/KeepRepository.cs
@@ -98,12 +98,12 @@ namespace keepr.Repositories
       throw new NotImplementedException();
     }
 
-    public string Delete(int id)
+    public string Delete(int id, string userId)
     {
       try
       {
-        string query = "DELETE FROM keeps WHERE id = @id;";
-        int changedRows = _db.Execute(query, new { id });
+        string query = "DELETE FROM keeps WHERE id = @id AND userId = @userId;";
+        int changedRows = _db.Execute(query, new { id, userId });
         if (changedRows < 1) throw new Exception("Invalid Id");
         return "Successfully Deleted";
       }
